+++ spec/www/spec/regexp-test.js
@@ -40,12 +40,12 @@ var mytests = function() {
 
       it(suiteName + 'Simple REGEXP test',
         function(done) {
-          if (isWP8) pending('NOT IMPLEMENTED for WP8');
-          if (isWindows) pending('NOT IMPLEMENTED for Windows');
-          // if (!isWebSql && !isWindows && isAndroid && /Android [1-4]/.test(navigator.userAgent)) pending('BROKEN for android.database (Android 1.x-4.x)');
-          if (!isWebSql && !isWindows && isAndroid) pending('NOT IMPLEMENTED for Android plugin');
+          if (isWP8) pending('NOT IMPLEMENTED for WP8 (plugin)');
+          if (isWindows) pending('NOT IMPLEMENTED for Windows (plugin)');
+          if (!isWebSql && !isWindows && isAndroid) pending('SKIP for Android plugin'); // TBD SKIP for Android plugin (for now)
+          if (!isWebSql && !isWindows && isAndroid) pending('SKIP: NOT IMPLEMENTED for Android plugin');
           if (isWebSql && !isAndroid && !isWindows && !isWP8) pending('SKIP for iOS (WebKit) Web SQL');
-          if (!isWebSql && !isAndroid && !isWindows && !isWP8) pending('NOT IMPLEMENTED for iOS/macOS');
+          if (!isWebSql && !isAndroid && !isWindows && !isWP8) pending('NOT IMPLEMENTED for iOS/macOS plugin');
 
           var db = openDatabase('simple-regexp-test.db', '1.0', 'test', DEFAULT_SIZE);